Changed a couple of decls (and casts) to make it (I hope) 64-bit compatible.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7169 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/packages/dbdrivers/cc/driver_manager_defs.h
@@ -74,7 +74,7 @@
 
 union xsb_value
 {
-  int i_val;
+  Integer i_val;
   double f_val;
   char* str_val;
 };
@@ -82,7 +82,7 @@ union xsb_value
 struct xsb_data
 {
   int type;
-  int length;
+  size_t length;
   union xsb_value* val;
 };
